Refactor Flash support.
- Remove autotools based build suport.
- Add npm flash build support.
- Move mod_fsp to flash/.
- Move swf/ to flash/swf/.
- Move policy servers to flash/.
- Update READMEs.
