--- Makefile.in
@@ -1,84 +1,0 @@
-# Makefile for Forge
-
-# Top-level build and dist dir
-BUILD_DIR=@FORGE_DIR@/build
-TOP_DIST_DIR=@FORGE_DIR@/dist
-DIST_DIR=$(TOP_DIST_DIR)/forge
-
-_FLASH := $(DIST_DIR)/SocketPool.swf
-ifeq (@BUILD_FLASH@,yes)
-FLASH := $(_FLASH)
-else
-ifeq (@USE_PRE_BUILT_FLASH@,yes)
-FLASH := $(_FLASH)
-endif
-endif
-JS_SOURCES := $(wildcard js/*.js)
-JS_DIST := $(JS_SOURCES:js/%.js=$(DIST_DIR)/%.js)
-JS_DIST_MIN := $(JS_DIST:%.js=%.min.js)
-TESTS_FORGE_LINK := @FORGE_DIR@/tests/forge
-
-# Whether or not to print commands as they are being executed, helpful for
-# debugging the build system.
-ifdef PRINT_COMMANDS
-PCMD=
-else
-PCMD=@
-endif
-
-.PHONY: all build-all update-all verbose clean verbose-commands
-
-# debug flags for flash build
-ifeq (@MXMLC_DEBUG_MODE@,yes)
-FLASH_FLAGS = \
-	-debug=true \
-	-define=CONFIG::debugging,true \
-	-define=CONFIG::release,false
-else
-FLASH_FLAGS = \
-	-debug=false \
-	-define=CONFIG::debugging,false \
-	-define=CONFIG::release,true
-endif
-
-all: $(BUILD_DIR) $(DIST_DIR) $(FLASH) $(JS_DIST) $(TESTS_FORGE_LINK)
-	@echo "forge build complete."
-
-build-all: all
-
-update-all:
-	@git pull && ./build-setup && make all
-
-$(BUILD_DIR):
-	$(PCMD) mkdir -p $@
-$(DIST_DIR):
-	$(PCMD) mkdir -p $@
-
-ifeq (@BUILD_FLASH@,yes)
-$(DIST_DIR)/SocketPool.swf: flash/SocketPool.as flash/PooledSocket.as flash/SocketEvent.as
-	@echo "Building $@..."
-	$(PCMD) @MXMLC@ $(FLASH_FLAGS) \
-		-load-config+=build-flash.xml \
-		-output=$@ $<
-else
-ifeq (@USE_PRE_BUILT_FLASH@,yes)
-$(DIST_DIR)/SocketPool.swf: @FORGE_DIR@/swf/SocketPool.swf
-	@echo "Copying pre-built $(@F)..."
-	$(PCMD) cp $< $@
-endif
-endif
-
-$(DIST_DIR)/%.js: js/%.js
-	@echo "Linking $@..."
-	$(PCMD) ln -sf $(realpath $<) $@
-
-$(TESTS_FORGE_LINK): $(DIST_DIR)
-	@echo "Linking $@..."
-	$(PCMD) ln -sf $(realpath $<) $@
-
-clean:
-	$(PCMD) rm -rf $(BUILD_DIR) $(TOP_DIST_DIR)
-	@echo "Removed all generated files."
-
-verbose-commands:
-	PRINT_COMMANDS=true $(MAKE) all
