--- build-flash.xml
@@ -1,7 +1,0 @@
-<flex-config>
-   <compiler>
-      <source-path>
-         <path-element>flash</path-element>
-      </source-path>
-   </compiler>
-</flex-config>
